worker/requirements.txt
aiokafka==0.12.0
async-timeout==4.0.3
asyncpg==0.30.0
certifi==2025.7.14
charset-normalizer==3.4.2
greenlet==3.2.3
idna==3.10
kafka-python==2.2.15
packaging==25.0
playwright==1.53.0
psycopg2==2.9.10
pyee==13.0.0
python-dotenv==1.1.1
requests==2.32.4
SQLAlchemy==2.0.41
typing_extensions==4.14.1
urllib3==2.5.0